refactor(nav): split NavGraph.kt into eleven per-feature graph extensions
NavGraph.kt was 1,420 lines: 69 screen imports, all 94 destinations. Every screen
in the app was reachable from one file, so no feature could move into its own
module — :app would have had to import from all eleven at once. This is the
critical path; PRs 13-21 are blocked on it. The file is now 281 lines holding the
NavHost, the navigation-suite shell, and eleven calls.

The route-to-feature mapping was derived from the source, not by hand: each
taggedComposable block matched against the screens it calls. Two things made that
non-trivial. Seven blocks call their screen fully-qualified rather than through an
import, so an import-only scan misses them. And a first attempt collapsed the 14
settings imports into one, because building the lookup with dict() on
(package, screen) pairs keeps only the last screen per package. With both fixed,
94 of 94 map to exactly one feature and the distribution matches SPEC.md's eleven
modules: quran 21, content 21, settings 16, tracker 11, prayer 8, about 7,
search 4, tools 2, calendar 2, onboarding 1, home 1.

The blocks turned out to be almost perfectly self-contained. Of the fourteen
values in NavGraph's scope they reference two: navController (all 94) and
analyticsContext (one, Route.Home).

Against #563, the graph functions take a NavController rather than
`onNavigate: (Route) -> Unit`. Raised on the issue before committing to it: 11 of
the 158 navigate calls in these blocks pass a NavOptionsBuilder — popUpTo,
launchSingleTop — which (Route) -> Unit cannot express, and flattening them would
change back-stack behaviour with no build error and no test failure, surfacing as
a wrong back button several screens later. A graph function IS navigation wiring,
so holding the controller is its job; the rule that matters is that a SCREEN must
not. NavControllerConfinementTest therefore exempts *Graph.kt and nothing else —
widening a guard added two commits earlier, so it also now asserts the eleven
graph files still exist, or the exemption could cover an empty set after a
re-merge.

EveryRouteIsRegisteredTest is the exit criterion #563 actually needs. NAV-03
compares TOTALS, so a route dropped in one graph while another is duplicated
leaves the count at 94 — precisely the shape a copy-paste across eleven new files
produces, and it fails as a blank screen at runtime rather than a build error.
The new test compares SETS in both directions and fails on duplicates
(Navigation Compose keeps the last registration, so the earlier block becomes
dead code that still reads as live). Demonstrated by doing exactly that to
QuranGraph: NAV-03 and NAV-05 both stayed green, both new assertions failed.

It caught its own documentation on the first run, reporting a Route.X that does
not exist — from the KDoc quoting taggedComposable<Route.X>. The identical trap
caught NAV-03 in PR 11. It blanks comments before scanning, as check_docs.py
learned to.

Verification: the extracted route set diffs byte-identical against the original —
none dropped, none duplicated. All ten NAV checks green across the eleven files.
:app 1077 tests (1082 - 12 for AmountFormattingTest moving to :core:ui, + 3 + 4
for the two new guards), one failure, the credential-gated DeviceStateCorpusTest.
:core:{navigation,ui,common}:check and :app:compileDebugAndroidTestKotlin green.

One trap worth recording: pruning imports by symbol name breaks `by` delegates.
Removing `import androidx.compose.runtime.getValue` broke eight sites, because the
symbol never appears literally in the source that needs it.

Docs updated in the same commit: NAVIGATION.md (the graph is eleven files, the
owning-file table, and where a new destination goes), ARCHITECTURE.md, CLAUDE.md
rule 6.

Closes #563. Stacked PR 12 of 22, epic #551.
