+++ app/src/main/java/com/arshadshah/nimaz/core/navigation/NavGraph.kt
@@ -3,7 +3,6 @@ package com.arshadshah.nimaz.core.navigation
 import android.app.Activity
 import android.content.Context
 import android.content.Intent
-import androidx.compose.animation.AnimatedContentScope
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material.icons.Icons
@@ -26,111 +25,37 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.res.stringResource
 import androidx.core.net.toUri
 import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
-import androidx.navigation.NavBackStackEntry
 import androidx.navigation.NavDestination.Companion.hasRoute
 import androidx.navigation.NavDestination.Companion.hierarchy
 import androidx.navigation.NavGraph.Companion.findStartDestination
-import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
-import androidx.navigation.toRoute
 import com.arshadshah.nimaz.core.monitoring.AppAnalytics
-import com.arshadshah.nimaz.core.share.ContentShareManager
-import com.arshadshah.nimaz.core.share.Shareables
 import com.arshadshah.nimaz.presentation.components.atoms.NimazIcon
-import com.arshadshah.nimaz.presentation.components.molecules.ShareAppSheet
-import com.arshadshah.nimaz.presentation.screens.about.AboutScreen
-import com.arshadshah.nimaz.presentation.screens.about.LicenseDetailScreen
-import com.arshadshah.nimaz.presentation.screens.about.LicensesScreen
-import com.arshadshah.nimaz.presentation.screens.adaptive.AdaptiveNamesScreen
-import com.arshadshah.nimaz.presentation.screens.adaptive.AdaptiveDuaScreen
-import com.arshadshah.nimaz.presentation.screens.adaptive.AdaptiveHadithScreen
-import com.arshadshah.nimaz.presentation.screens.adaptive.AdaptiveKhatamScreen
-import com.arshadshah.nimaz.presentation.screens.adaptive.AdaptiveMoreScreen
-import com.arshadshah.nimaz.presentation.screens.adaptive.AdaptiveQuranScreen
-import com.arshadshah.nimaz.presentation.screens.adaptive.AdaptiveSettingsScreen
-import com.arshadshah.nimaz.presentation.screens.asma.AsmaUlHusnaDetailScreen
-import com.arshadshah.nimaz.presentation.screens.names.FavouritesScreen
-import com.arshadshah.nimaz.core.navigation.NamesTab
-import com.arshadshah.nimaz.presentation.screens.asmaunnabi.AsmaUnNabiDetailScreen
-import com.arshadshah.nimaz.presentation.screens.bookmarks.SavedScreen
-import com.arshadshah.nimaz.presentation.screens.calendar.IslamicCalendarScreen
-import com.arshadshah.nimaz.presentation.screens.dua.DuaCategoryScreen
-import com.arshadshah.nimaz.presentation.screens.dua.DuaOccasionScreen
-import com.arshadshah.nimaz.presentation.screens.dua.DuaReaderScreen
-import com.arshadshah.nimaz.presentation.screens.dua.DuaSettingsScreen
-import com.arshadshah.nimaz.presentation.screens.dua.DuasCollectionScreen
-import com.arshadshah.nimaz.presentation.screens.fasting.FastTrackerScreen
-import com.arshadshah.nimaz.presentation.screens.fasting.MakeupFastsScreen
-import com.arshadshah.nimaz.presentation.screens.hadith.HadithChaptersScreen
-import com.arshadshah.nimaz.presentation.screens.hadith.HadithReaderScreen
-import com.arshadshah.nimaz.presentation.screens.hadith.HadithSettingsScreen
-import com.arshadshah.nimaz.presentation.screens.home.HomeScreen
-import com.arshadshah.nimaz.presentation.screens.khatam.KhatamDetailScreen
-import com.arshadshah.nimaz.presentation.screens.khatam.KhatamFormScreen
-import com.arshadshah.nimaz.presentation.screens.onboarding.OnboardingScreen
-import com.arshadshah.nimaz.presentation.screens.prayer.MonthlyPrayerTimesScreen
-import com.arshadshah.nimaz.presentation.screens.prayer.PrayerStatsScreen
-import com.arshadshah.nimaz.presentation.screens.prayer.PrayerTimesScreen
-import com.arshadshah.nimaz.presentation.screens.prayer.PrayerTrackerScreen
-import com.arshadshah.nimaz.presentation.screens.prayer.QadaPrayersScreen
-import com.arshadshah.nimaz.presentation.screens.prophets.ProphetDetailScreen
-import com.arshadshah.nimaz.presentation.screens.qaida.QaidaHomeScreen
-import com.arshadshah.nimaz.presentation.screens.qaida.QaidaLettersScreen
-import com.arshadshah.nimaz.presentation.screens.qaida.QaidaReaderScreen
-import com.arshadshah.nimaz.presentation.screens.qibla.QiblaScreen
-import com.arshadshah.nimaz.presentation.screens.quran.QuranBrowseScreen
-import com.arshadshah.nimaz.presentation.screens.quran.QuranReaderScreen
-import com.arshadshah.nimaz.presentation.screens.quran.SelectReciterScreen
-import com.arshadshah.nimaz.presentation.screens.quran.SelectTranslationScreen
-import com.arshadshah.nimaz.domain.model.DuaOccasion
-import com.arshadshah.nimaz.domain.model.NameCatalog
-import com.arshadshah.nimaz.domain.model.HadithGrade
-import com.arshadshah.nimaz.domain.model.TopicTree
-import com.arshadshah.nimaz.presentation.screens.quran.QuranTopicDetailScreen
-import com.arshadshah.nimaz.presentation.screens.quran.QuranTopicsScreen
-import com.arshadshah.nimaz.presentation.screens.quran.SurahBackgroundScreen
-import com.arshadshah.nimaz.presentation.screens.quran.SurahPassagesScreen
-import com.arshadshah.nimaz.presentation.screens.quran.SurahSubjectsScreen
-import com.arshadshah.nimaz.presentation.screens.quran.TafseerChaptersScreen
-import com.arshadshah.nimaz.presentation.screens.quran.TafseerScreen
-import com.arshadshah.nimaz.presentation.screens.search.SearchScreen
-import com.arshadshah.nimaz.presentation.screens.settings.AppearanceSettingsScreen
-import com.arshadshah.nimaz.presentation.screens.settings.LanguageScreen
-import com.arshadshah.nimaz.presentation.screens.settings.LocationScreen
-import com.arshadshah.nimaz.presentation.screens.settings.NotificationSettingsScreen
-import com.arshadshah.nimaz.presentation.screens.settings.NotificationSoundScreen
-import com.arshadshah.nimaz.presentation.screens.settings.NotificationDiagnosticsScreen
-import com.arshadshah.nimaz.presentation.screens.settings.NotificationWeeklyScreen
-import com.arshadshah.nimaz.presentation.screens.settings.PrayerNotificationsScreen
-import com.arshadshah.nimaz.presentation.screens.settings.WorshipRemindersScreen
-import com.arshadshah.nimaz.presentation.screens.settings.PrayerSettingsScreen
-import com.arshadshah.nimaz.presentation.screens.settings.QuranSettingsScreen
-import com.arshadshah.nimaz.presentation.screens.settings.SearchSettingsScreen
-import com.arshadshah.nimaz.presentation.screens.settings.ZakatSettingsScreen
-import com.arshadshah.nimaz.presentation.screens.settings.WidgetsScreen
-import com.arshadshah.nimaz.presentation.screens.tasbih.TasbihScreen
-import com.arshadshah.nimaz.presentation.screens.zakat.ZakatCalculatorScreen
-import com.arshadshah.nimaz.presentation.screens.zakat.ZakatHistoryScreen
+import com.arshadshah.nimaz.presentation.screens.about.aboutGraph
+import com.arshadshah.nimaz.presentation.screens.calendar.calendarGraph
+import com.arshadshah.nimaz.presentation.screens.content.contentGraph
+import com.arshadshah.nimaz.presentation.screens.home.homeGraph
+import com.arshadshah.nimaz.presentation.screens.onboarding.onboardingGraph
+import com.arshadshah.nimaz.presentation.screens.prayer.prayerGraph
+import com.arshadshah.nimaz.presentation.screens.quran.quranGraph
+import com.arshadshah.nimaz.presentation.screens.search.searchGraph
+import com.arshadshah.nimaz.presentation.screens.settings.settingsGraph
+import com.arshadshah.nimaz.presentation.screens.tools.toolsGraph
+import com.arshadshah.nimaz.presentation.screens.tracker.trackerGraph
 import com.arshadshah.nimaz.presentation.theme.isTablet
 import com.arshadshah.nimaz.presentation.viewmodel.onboarding.OnboardingViewModel
-import com.arshadshah.nimaz.presentation.viewmodel.search.SearchFilter
-import com.google.android.play.core.review.ReviewManagerFactory
-import kotlinx.coroutines.launch
 import kotlin.system.exitProcess
-import com.arshadshah.nimaz.presentation.screens.worship.NightWorshipScreen
+import kotlinx.coroutines.launch
 
 @Composable
 fun NavGraph(
@@ -313,1086 +238,20 @@ fun NavGraph(
             navController = navController,
             startDestination = startDestination,
         ) {
-            // Onboarding
-            taggedComposable<Route.Onboarding>(ScreenTags.Onboarding) {
-                OnboardingScreen(
-                    onComplete = {
-                        navController.navigate(Route.Home) {
-                            popUpTo(Route.Onboarding) { inclusive = true }
-                        }
-                    }
-                )
-            }
-
-            // Main screens
-            taggedComposable<Route.Home>(ScreenTags.Home) {
-                HomeScreen(
-                    onNavigateToAlKahf = { navController.navigate(Route.QuranReader(surahNumber = 18)) },
-                    onNavigateToHadith = { navController.navigate(Route.HadithHome) },
-                    onNavigateToDua = { duaId -> navController.navigate(Route.DuaReader(duaId)) },
-                    onNavigateToTasbih = { navController.navigate(Route.TasbihHome) },
-                    onNavigateToCalendar = { navController.navigate(Route.IslamicCalendar) },
-                    onNavigateToFasting = { navController.navigate(Route.FastingHome) },
-                    onNavigateToZakat = { navController.navigate(Route.ZakatCalculator) },
-                    onNavigateToPrayerTracker = { navController.navigate(Route.PrayerTracker) },
-                    onNavigateToSettings = { navController.navigate(Route.Settings) },
-                    onNavigateToPrayerSettings = { navController.navigate(Route.SettingsPrayerCalculation) },
-                    onNavigateToPrayerTimes = { navController.navigate(Route.PrayerTimes) },
-                    onOpenHadith = { hadithId -> navController.navigate(Route.HadithReader(hadithId)) },
-                    // Every worship card now leads somewhere; the mapping is a pure function so
-                    // it is asserted in a JVM test rather than only through the UI.
-                    onOpenWorship = { type -> navController.navigate(worshipCardDestination(type)) },
-                    onOpenAnnouncementRoute = { key ->
-                        if (key.startsWith("https://")) {
-                            runCatching {
-                                analyticsContext.startActivity(
-                                    Intent(Intent.ACTION_VIEW, key.toUri())
-                                )
-                            }
-                        } else {
-                            announcementRoute(key)?.let { navController.navigate(it) }
-                        }
-                    }
-                )
-            }
-
-            taggedComposable<Route.Quran>(ScreenTags.Quran) {
-                AdaptiveQuranScreen(
-                    onNavigate = { navController.navigate(it) },
-                    onBack = { navController.popBackStack() },
-                    onNavigateToSearch = { navController.navigate(Route.GlobalSearch) },
-                    onNavigateToTopics = { navController.navigate(Route.QuranTopics) },
-                    onNavigateToBrowse = { navController.navigate(Route.QuranBrowse()) },
-                    onNavigateToSaved = { navController.navigate(Route.QuranSaved) },
-                    onNavigateToSettings = { navController.navigate(Route.SettingsQuran) },
-                    onNavigateToKhatam = { navController.navigate(Route.KhatamList) },
-                )
-            }
-
-            taggedComposable<Route.QuranBrowse>(ScreenTags.QuranBrowse) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.QuranBrowse>()
-                QuranBrowseScreen(
-                    initialInfoForSurah = args.infoForSurah,
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToSurah = { surah ->
-                        navController.navigate(Route.QuranReader(surah))
-                    },
-                    onNavigateToJuz = { juz -> navController.navigate(Route.QuranJuz(juz)) },
-                    onNavigateToPage = { page -> navController.navigate(Route.QuranPage(page)) },
-                    onOpenBackground = { surah ->
-                        navController.navigate(Route.SurahBackground(surah))
-                    },
-                    onOpenPassages = { surah ->
-                        navController.navigate(Route.SurahPassages(surah))
-                    },
-                    onOpenSubjects = { surah ->
-                        navController.navigate(Route.SurahSubjects(surah))
-                    },
-                )
-            }
-
-            taggedComposable<Route.Tasbih>(ScreenTags.Tasbih) {
-                TasbihScreen(
-                    onNavigateToHistory = { navController.navigate(Route.TasbihHistory) },
-                    onNavigateToChooseDhikr = { navController.navigate(Route.TasbihPresets) },
-                    onNavigateToAddPreset = { navController.navigate(Route.TasbihAddPreset()) },
-                    onNavigateToSettings = { navController.navigate(Route.Settings) }
-                )
-            }
-
-            taggedComposable<Route.QiblaNav>(ScreenTags.QiblaNav) {
-                QiblaScreen(
-                    onNavigateBack = { navController.popBackStack() }
-                )
-            }
-
-            // Qaida (children's Arabic reader) screens
-            taggedComposable<Route.QaidaHome>(ScreenTags.QaidaHome) {
-                QaidaHomeScreen(
-                    onNavigateBack = { navController.popBackStack() },
-                    onOpenLesson = { lessonId -> navController.navigate(Route.QaidaReader(lessonId)) },
-                    onOpenLetters = { navController.navigate(Route.QaidaLetters) },
-                )
-            }
-
-            taggedComposable<Route.QaidaReader>(ScreenTags.QaidaReader) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.QaidaReader>()
-                QaidaReaderScreen(
-                    lessonId = args.lessonId,
-                    onNavigateBack = { navController.popBackStack() },
-                )
-            }
-
-            taggedComposable<Route.QaidaLetters>(ScreenTags.QaidaLetters) {
-                QaidaLettersScreen(
-                    onNavigateBack = { navController.popBackStack() },
-                )
-            }
-
-            taggedComposable<Route.More>(ScreenTags.More) {
-                val context = androidx.compose.ui.platform.LocalContext.current
-                AdaptiveMoreScreen(
-                    onNavigate = { navController.navigate(it) },
-                    onRestartApp = { restartApp(context) },
-                )
-            }
-
-            // Quran screens
-            taggedComposable<Route.QuranReader>(ScreenTags.QuranReader) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.QuranReader>()
-                QuranReaderScreen(
-                    surahNumber = args.surahNumber,
-                    initialAyahNumber = args.ayahNumber,
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToQuranSettings = { navController.navigate(Route.SettingsQuran) },
-                    onNavigateToTafseer = { surah, ayah ->
-                        navController.navigate(Route.Tafseer(surah, ayah))
-                    },
-                    onNavigateToPassages = { surah, ayah ->
-                        navController.navigate(Route.SurahPassages(surah, ayah))
-                    },
-                    onNavigateToSubjects = { surah ->
-                        navController.navigate(
-                            if (surah != null) Route.SurahSubjects(surah) else Route.QuranTopics
-                        )
-                    },
-                    onNavigateToReciters = { navController.navigate(Route.SelectReciter) },
-                    onNavigateToNextSurah = { nextSurah ->
-                        navController.navigate(Route.QuranReader(nextSurah)) {
-                            popUpTo<Route.QuranReader> { inclusive = true }
-                        }
-                    }
-                )
-            }
-
-            taggedComposable<Route.TafseerChapters>(ScreenTags.TafseerChapters) {
-                TafseerChaptersScreen(
-                    onNavigateBack = { navController.popBackStack() },
-                    onOpenTafseer = { surah, ayah ->
-                        navController.navigate(
-                            Route.Tafseer(
-                                surahNumber = surah,
-                                ayahNumber = ayah
-                            )
-                        )
-                    }
-                )
-            }
-
-            taggedComposable<Route.Tafseer>(ScreenTags.Tafseer) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.Tafseer>()
-                TafseerScreen(
-                    surahNumber = args.surahNumber,
-                    ayahNumber = args.ayahNumber,
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToTopic = { topicId ->
-                        navController.navigate(Route.QuranTopicDetail(topicId))
-                    }
-                )
-            }
-
-            taggedComposable<Route.SurahBackground>(ScreenTags.SurahBackground) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.SurahBackground>()
-                SurahBackgroundScreen(
-                    surahNumber = args.surahNumber,
-                    onNavigateBack = { navController.popBackStack() },
-                    onOpenAyah = { surah, ayah ->
-                        navController.navigate(Route.QuranReader(surah, ayah))
-                    },
-                    onOpenTopic = { topicId ->
-                        navController.navigate(Route.QuranTopicDetail(topicId))
-                    }
-                )
-            }
-
-            taggedComposable<Route.SurahPassages>(ScreenTags.SurahPassages) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.SurahPassages>()
-                SurahPassagesScreen(
-                    surahNumber = args.surahNumber,
-                    currentAyah = args.currentAyah,
-                    onNavigateBack = { navController.popBackStack() },
-                    onOpenAyah = { surah, ayah ->
-                        navController.navigate(Route.QuranReader(surah, ayah))
-                    }
-                )
-            }
-
-            taggedComposable<Route.SurahSubjects>(ScreenTags.SurahSubjects) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.SurahSubjects>()
-                SurahSubjectsScreen(
-                    surahNumber = args.surahNumber,
-                    onNavigateBack = { navController.popBackStack() },
-                    onOpenTopic = { topicId, tree ->
-                        navController.navigate(
-                            Route.QuranTopicDetail(topicId, tree.wire, args.surahNumber)
-                        )
-                    },
-                    onBrowseAllSubjects = { navController.navigate(Route.QuranTopics) }
-                )
-            }
-
-            taggedComposable<Route.QuranTopics>(ScreenTags.QuranTopics) {
-                QuranTopicsScreen(
-                    onNavigateBack = { navController.popBackStack() },
-                    onOpenTopic = { topicId, tree ->
-                        navController.navigate(Route.QuranTopicDetail(topicId, tree.wire))
-                    }
-                )
-            }
-
-            taggedComposable<Route.QuranTopicDetail>(ScreenTags.QuranTopicDetail) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.QuranTopicDetail>()
-                QuranTopicDetailScreen(
-                    topicId = args.topicId,
-                    tree = TopicTree.fromWire(args.tree),
-                    fromSurah = args.fromSurah,
-                    onNavigateBack = { navController.popBackStack() },
-                    onOpenAyah = { surah, ayah ->
-                        navController.navigate(Route.QuranReader(surah, ayah))
-                    },
-                    // The surah travels with every lateral move — a subtopic, a related
-                    // subject, a cross-link in the prose. Dropping it one hop in would put the
-                    // reader back where this whole change started, one screen later.
-                    onOpenTopic = { topicId, tree ->
-                        navController.navigate(
-                            Route.QuranTopicDetail(topicId, tree.wire, args.fromSurah)
-                        )
-                    }
-                )
-            }
-
-            taggedComposable<Route.QuranPage>(ScreenTags.QuranPage) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.QuranPage>()
-                QuranReaderScreen(
-                    pageNumber = args.pageNumber,
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToQuranSettings = { navController.navigate(Route.SettingsQuran) },
-                    onNavigateToTafseer = { surah, ayah ->
-                        navController.navigate(Route.Tafseer(surah, ayah))
-                    },
-                    onNavigateToPassages = { surah, ayah ->
-                        navController.navigate(Route.SurahPassages(surah, ayah))
-                    },
-                    onNavigateToSubjects = { surah ->
-                        navController.navigate(
-                            if (surah != null) Route.SurahSubjects(surah) else Route.QuranTopics
-                        )
-                    },
-                    onNavigateToReciters = { navController.navigate(Route.SelectReciter) }
-                )
-            }
-
-            taggedComposable<Route.QuranJuz>(ScreenTags.QuranJuz) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.QuranJuz>()
-                QuranReaderScreen(
-                    juzNumber = args.juzNumber,
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToQuranSettings = { navController.navigate(Route.SettingsQuran) },
-                    onNavigateToTafseer = { surah, ayah ->
-                        navController.navigate(Route.Tafseer(surah, ayah))
-                    },
-                    onNavigateToPassages = { surah, ayah ->
-                        navController.navigate(Route.SurahPassages(surah, ayah))
-                    },
-                    onNavigateToSubjects = { surah ->
-                        navController.navigate(
-                            if (surah != null) Route.SurahSubjects(surah) else Route.QuranTopics
-                        )
-                    },
-                    onNavigateToReciters = { navController.navigate(Route.SelectReciter) }
-                )
-            }
-
-            taggedComposable<Route.QuranSearch>(ScreenTags.QuranSearch) {
-                SearchScreen(
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToQuranAyah = { surah, ayah ->
-                        navController.navigate(Route.QuranReader(surah, ayah))
-                    },
-                    onNavigateToSurah = { surah ->
-                        navController.navigate(Route.QuranReader(surah))
-                    },
-                    onNavigateToHadith = { bookId, hadithId ->
-                        navController.navigate(Route.HadithReader(hadithId))
-                    },
-                    onNavigateToDua = { duaId ->
-                        navController.navigate(Route.DuaReader(duaId))
-                    },
-                    onNavigateToName = { catalog, id ->
-                        navController.navigate(
-                            when (catalog) {
-                                NameCatalog.ASMA_UL_HUSNA -> Route.AsmaUlHusnaDetail(id)
-                                NameCatalog.ASMA_UN_NABI -> Route.AsmaUnNabiDetail(id)
-                                NameCatalog.PROPHETS -> Route.ProphetDetail(id)
-                            }
-                        )
-                    },
-                )
-            }
-
-            taggedComposable<Route.QuranSaved>(ScreenTags.QuranSaved) {
-                SavedScreen(
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToQuranAyah = { surah, ayah ->
-                        navController.navigate(Route.QuranReader(surah, ayah))
-                    },
-                    onNavigateToHadith = { bookId, hadithNumber ->
-                        navController.navigate(Route.HadithByNumber(bookId, hadithNumber))
-                    },
-                    onNavigateToDua = { duaId ->
-                        navController.navigate(Route.DuaReader(duaId))
-                    },
-                )
-            }
-
-            // Hadith screens
-            taggedComposable<Route.HadithHome>(ScreenTags.HadithHome) {
-                AdaptiveHadithScreen(
-                    onNavigate = { navController.navigate(it) },
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToSearch = { navController.navigate(Route.HadithSearch) },
-                    onNavigateToBookmarks = { navController.navigate(Route.HadithBookmarks) },
-                    onNavigateToGrade = { grade ->
-                        navController.navigate(Route.HadithByGrade(grade.name))
-                    },
-                )
-            }
-
-            taggedComposable<Route.HadithBook>(ScreenTags.HadithBook) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.HadithBook>()
-                HadithChaptersScreen(
-                    bookId = args.bookId,
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToChapter = { bookId, chapterId ->
-                        navController.navigate(Route.HadithChapter(bookId, chapterId))
-                    }
-                )
-            }
-
-            taggedComposable<Route.HadithChapter>(ScreenTags.HadithChapter) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.HadithChapter>()
-                HadithReaderScreen(
-                    bookId = args.bookId,
-                    chapterId = args.chapterId,
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToSettings = { navController.navigate(Route.HadithSettings) }
-                )
-            }
-
-            taggedComposable<Route.HadithReader>(ScreenTags.HadithReader) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.HadithReader>()
-                HadithReaderScreen(
-                    bookId = "",
-                    chapterId = args.hadithId,
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToSettings = { navController.navigate(Route.HadithSettings) }
-                )
-            }
-
-            taggedComposable<Route.HadithByNumber>(ScreenTags.HadithByNumber) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.HadithByNumber>()
-                HadithReaderScreen(
-                    bookId = args.bookId,
-                    chapterId = "",
-                    hadithNumber = args.hadithNumber,
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToSettings = { navController.navigate(Route.HadithSettings) }
-                )
-            }
-
-            taggedComposable<Route.HadithByGrade>(ScreenTags.HadithByGrade) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.HadithByGrade>()
-                HadithReaderScreen(
-                    bookId = "",
-                    chapterId = "",
-                    grade = HadithGrade.entries.firstOrNull { it.name == args.grade },
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToSettings = { navController.navigate(Route.HadithSettings) }
-                )
-            }
-
-            taggedComposable<Route.HadithSettings>(ScreenTags.HadithSettings) {
-                HadithSettingsScreen(
-                    onNavigateBack = { navController.popBackStack() }
-                )
-            }
-
-            taggedComposable<Route.HadithSearch>(ScreenTags.HadithSearch) {
-                SearchScreen(
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToQuranAyah = { surah, ayah ->
-                        navController.navigate(Route.QuranReader(surah, ayah))
-                    },
-                    onNavigateToSurah = { surah ->
-                        navController.navigate(Route.QuranReader(surah))
-                    },
-                    onNavigateToHadith = { bookId, hadithId ->
-                        navController.navigate(Route.HadithReader(hadithId))
-                    },
-                    onNavigateToDua = { duaId ->
-                        navController.navigate(Route.DuaReader(duaId))
-                    },
-                    onNavigateToName = { catalog, id ->
-                        navController.navigate(
-                            when (catalog) {
-                                NameCatalog.ASMA_UL_HUSNA -> Route.AsmaUlHusnaDetail(id)
-                                NameCatalog.ASMA_UN_NABI -> Route.AsmaUnNabiDetail(id)
-                                NameCatalog.PROPHETS -> Route.ProphetDetail(id)
-                            }
-                        )
-                    },
-                )
-            }
-
-            taggedComposable<Route.HadithBookmarks>(ScreenTags.HadithBookmarks) {
-                SavedScreen(
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToQuranAyah = { surah, ayah ->
-                        navController.navigate(Route.QuranReader(surah, ayah))
-                    },
-                    onNavigateToHadith = { bookId, hadithNumber ->
-                        navController.navigate(Route.HadithByNumber(bookId, hadithNumber))
-                    },
-                    onNavigateToDua = { duaId ->
-                        navController.navigate(Route.DuaReader(duaId))
-                    },
-                )
-            }
-
-            // Dua screens
-            taggedComposable<Route.DuaHome>(ScreenTags.DuaHome) {
-                AdaptiveDuaScreen(
-                    onNavigate = { navController.navigate(it) },
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToBookmarks = { navController.navigate(Route.AllBookmarks) },
-                    onNavigateToSearch = { navController.navigate(Route.DuaSearch) },
-                )
-            }
-
-            taggedComposable<Route.DuaCategory>(ScreenTags.DuaCategory) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.DuaCategory>()
-                DuaCategoryScreen(
-                    categoryId = args.categoryId,
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToDua = { duaId ->
-                        navController.navigate(Route.DuaReader(duaId))
-                    },
-                    onNavigateToOccasion = { occasion ->
-                        navController.navigate(Route.DuaOccasion(occasion.name))
-                    }
-                )
-            }
-
-            taggedComposable<Route.DuaOccasion>(ScreenTags.DuaOccasion) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.DuaOccasion>()
-                DuaOccasionScreen(
-                    occasion = DuaOccasion.entries.firstOrNull { it.name == args.occasion }
-                        ?: DuaOccasion.GENERAL,
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToDua = { duaId ->
-                        navController.navigate(Route.DuaReader(duaId))
-                    },
-                )
-            }
-
-            taggedComposable<Route.DuaReader>(ScreenTags.DuaReader) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.DuaReader>()
-                DuaReaderScreen(
-                    duaId = args.duaId,
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToSettings = { navController.navigate(Route.DuaSettings) }
-                )
-            }
-
-            taggedComposable<Route.DuaSettings>(ScreenTags.DuaSettings) {
-                DuaSettingsScreen(
-                    onNavigateBack = { navController.popBackStack() }
-                )
-            }
-
-            taggedComposable<Route.DuaFavorites>(ScreenTags.DuaFavorites) {
-                DuasCollectionScreen(
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToCategory = { categoryId ->
-                        navController.navigate(Route.DuaCategory(categoryId))
-                    },
-                    onNavigateToBookmarks = { navController.navigate(Route.AllBookmarks) },
-                    onNavigateToSearch = { navController.navigate(Route.DuaSearch) }
-                )
-            }
-
-            taggedComposable<Route.DuaSearch>(ScreenTags.DuaSearch) {
-                SearchScreen(
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToQuranAyah = { surah, ayah ->
-                        navController.navigate(Route.QuranReader(surah, ayah))
-                    },
-                    onNavigateToSurah = { surah ->
-                        navController.navigate(Route.QuranReader(surah))
-                    },
-                    onNavigateToHadith = { bookId, hadithId ->
-                        navController.navigate(Route.HadithReader(hadithId))
-                    },
-                    onNavigateToDua = { duaId ->
-                        navController.navigate(Route.DuaReader(duaId))
-                    },
-                    onNavigateToName = { catalog, id ->
-                        navController.navigate(
-                            when (catalog) {
-                                NameCatalog.ASMA_UL_HUSNA -> Route.AsmaUlHusnaDetail(id)
-                                NameCatalog.ASMA_UN_NABI -> Route.AsmaUnNabiDetail(id)
-                                NameCatalog.PROPHETS -> Route.ProphetDetail(id)
-                            }
-                        )
-                    },
-                    initialFilter = SearchFilter.DUA
-                )
-            }
-
-            // Prayer screens
-            taggedComposable<Route.PrayerTimes>(ScreenTags.PrayerTimes) {
-                PrayerTimesScreen(
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToSettings = { navController.navigate(Route.SettingsPrayerCalculation) }
-                )
-            }
-
-            taggedComposable<Route.PrayerTracker>(ScreenTags.PrayerTracker) {
-                PrayerTrackerScreen(
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToStats = { navController.navigate(Route.PrayerStats) },
-                    onNavigateToQada = { navController.navigate(Route.QadaPrayers) },
-                )
-            }
-
-            taggedComposable<Route.PrayerStats>(ScreenTags.PrayerStats) {
-                PrayerStatsScreen(
-                    onNavigateBack = { navController.popBackStack() }
-                )
-            }
-
-            taggedComposable<Route.QadaPrayers>(ScreenTags.QadaPrayers) {
-                QadaPrayersScreen(onNavigateBack = { navController.popBackStack() })
-            }
-
-            taggedComposable<Route.MonthlyPrayerTimes>(ScreenTags.MonthlyPrayerTimes) {
-                MonthlyPrayerTimesScreen(
-                    onNavigateBack = { navController.popBackStack() }
-                )
-            }
-
-            // Fasting screens
-            taggedComposable<Route.FastingHome>(ScreenTags.FastingHome) {
-                FastTrackerScreen(
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToCalendar = { navController.navigate(Route.IslamicCalendar) },
-                    onNavigateToMakeup = { navController.navigate(Route.MakeupFasts) },
-                )
-            }
-
-            taggedComposable<Route.FastingTracker>(ScreenTags.FastingTracker) {
-                FastTrackerScreen(
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToCalendar = { navController.navigate(Route.IslamicCalendar) },
-                    onNavigateToMakeup = { navController.navigate(Route.MakeupFasts) },
-                )
-            }
-
-            taggedComposable<Route.NightWorship>(ScreenTags.NightWorship) {
-                NightWorshipScreen(
-                    onNavigateBack = { navController.popBackStack() },
-                    onOpenSurah = { surah: Int -> navController.navigate(Route.QuranReader(surah)) },
-                    onOpenDuaCategory = { id: String -> navController.navigate(Route.DuaCategory(id)) },
-                    onOpenHadith = { id: String -> navController.navigate(Route.HadithReader(id)) },
-                )
-            }
-
-            taggedComposable<Route.FastingStats>(ScreenTags.FastingStats) {
-                FastTrackerScreen(
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToCalendar = { navController.navigate(Route.IslamicCalendar) },
-                    onNavigateToMakeup = { navController.navigate(Route.MakeupFasts) },
-                )
-            }
-
-            taggedComposable<Route.MakeupFasts>(ScreenTags.MakeupFasts) {
-                MakeupFastsScreen(
-                    onNavigateBack = { navController.popBackStack() }
-                )
-            }
-
-            // Tasbih screens
-            taggedComposable<Route.TasbihHome>(ScreenTags.TasbihHome) {
-                TasbihScreen(
-                    onNavigateToHistory = { navController.navigate(Route.TasbihHistory) },
-                    onNavigateToChooseDhikr = { navController.navigate(Route.TasbihPresets) },
-                    onNavigateToAddPreset = { navController.navigate(Route.TasbihAddPreset()) },
-                    onNavigateToSettings = { navController.navigate(Route.Settings) }
-                )
-            }
-
-            taggedComposable<Route.TasbihCounter>(ScreenTags.TasbihCounterScreen) { backStackEntry ->
-                backStackEntry.toRoute<Route.TasbihCounter>()
-                TasbihScreen(
-                    onNavigateToHistory = { navController.navigate(Route.TasbihStats) },
-                    onNavigateToChooseDhikr = { navController.navigate(Route.TasbihPresets) },
-                    onNavigateToSettings = { navController.navigate(Route.Settings) }
-                )
-            }
-
-            taggedComposable<Route.TasbihPresets>(ScreenTags.TasbihPresets) {
-                com.arshadshah.nimaz.presentation.screens.tasbih.ChooseDhikrScreen(
-                    onBack = { navController.popBackStack() },
-                    onNavigateToAddPreset = { navController.navigate(Route.TasbihAddPreset()) },
-                    onEditPreset = { presetId ->
-                        navController.navigate(Route.TasbihAddPreset(presetId))
-                    }
-                )
-            }
-
-            taggedComposable<Route.TasbihStats>(ScreenTags.TasbihStats) {
-                TasbihScreen(
-                    onNavigateToHistory = { },
-                    onNavigateToSettings = { navController.navigate(Route.Settings) }
-                )
-            }
-
-            taggedComposable<Route.TasbihHistory>(ScreenTags.TasbihHistory) {
-                com.arshadshah.nimaz.presentation.screens.tasbih.TasbihHistoryScreen(
-                    onNavigateBack = { navController.popBackStack() }
-                )
-            }
-
-            taggedComposable<Route.TasbihAddPreset>(ScreenTags.TasbihAddPreset) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.TasbihAddPreset>()
-                com.arshadshah.nimaz.presentation.screens.tasbih.AddPresetScreen(
-                    presetId = args.presetId,
-                    onNavigateBack = { navController.popBackStack() }
-                )
-            }
-
-            // Zakat screens
-            taggedComposable<Route.ZakatCalculator>(ScreenTags.ZakatCalculator) {
-                ZakatCalculatorScreen(
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToHistory = { navController.navigate(Route.ZakatHistory) },
-                    onNavigateToSettings = { navController.navigate(Route.SettingsZakat) }
-                )
-            }
-
-            taggedComposable<Route.SettingsZakat>(ScreenTags.SettingsZakat) {
-                ZakatSettingsScreen(
-                    onNavigateBack = { navController.popBackStack() }
-                )
-            }
-
-            taggedComposable<Route.ZakatHistory>(ScreenTags.ZakatHistory) {
-                ZakatHistoryScreen(
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToCalculator = { navController.navigate(Route.ZakatCalculator) }
-                )
-            }
-
-            // Qibla
-            taggedComposable<Route.Qibla>(ScreenTags.Qibla) {
-                QiblaScreen(
-                    onNavigateBack = { navController.popBackStack() }
-                )
-            }
-
-            // Islamic Calendar
-            taggedComposable<Route.IslamicCalendar>(ScreenTags.IslamicCalendar) {
-                IslamicCalendarScreen(
-                    onNavigateBack = { navController.popBackStack() }
-                )
-            }
-
-            taggedComposable<Route.IslamicMonth>(ScreenTags.IslamicMonth) { backStackEntry ->
-                backStackEntry.toRoute<Route.IslamicMonth>()
-                IslamicCalendarScreen(
-                    onNavigateBack = { navController.popBackStack() }
-                )
-            }
-
-            // Settings
-            taggedComposable<Route.Settings>(ScreenTags.Settings) {
-                val context = androidx.compose.ui.platform.LocalContext.current
-                AdaptiveSettingsScreen(
-                    onNavigate = { navController.navigate(it) },
-                    onBack = { navController.popBackStack() },
-                    onRestartApp = { restartApp(context) },
-                )
-            }
-
-            taggedComposable<Route.SettingsPrayerCalculation>(ScreenTags.SettingsPrayerCalculation) {
-                PrayerSettingsScreen(
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToNotifications = { navController.navigate(Route.SettingsNotifications) }
-                )
-            }
-
-            taggedComposable<Route.SettingsNotifications>(ScreenTags.SettingsNotifications) {
-                NotificationSettingsScreen(
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToPrayers = { navController.navigate(Route.SettingsNotificationsPrayers) },
-                    onNavigateToWorshipReminders = { navController.navigate(Route.SettingsWorshipReminders) },
-                    onNavigateToWeekly = { navController.navigate(Route.SettingsNotificationsWeekly) },
-                    onNavigateToSound = { navController.navigate(Route.SettingsNotificationsSound) },
-                    onNavigateToDiagnostics = { navController.navigate(Route.SettingsNotificationsDiagnostics) },
-                )
-            }
-
-            taggedComposable<Route.SettingsWorshipReminders>(
-                ScreenTags.SettingsWorshipReminders
-            ) {
-                WorshipRemindersScreen(onNavigateBack = { navController.popBackStack() })
-            }
-            taggedComposable<Route.SettingsNotificationsPrayers>(
-                ScreenTags.SettingsNotificationsPrayers
-            ) {
-                PrayerNotificationsScreen(onNavigateBack = { navController.popBackStack() })
-            }
-            taggedComposable<Route.SettingsNotificationsWeekly>(
-                ScreenTags.SettingsNotificationsWeekly
-            ) {
-                NotificationWeeklyScreen(onNavigateBack = { navController.popBackStack() })
-            }
-            taggedComposable<Route.SettingsNotificationsSound>(
-                ScreenTags.SettingsNotificationsSound
-            ) {
-                NotificationSoundScreen(onNavigateBack = { navController.popBackStack() })
-            }
-            taggedComposable<Route.SettingsNotificationsDiagnostics>(
-                ScreenTags.SettingsNotificationsDiagnostics
-            ) {
-                NotificationDiagnosticsScreen(onNavigateBack = { navController.popBackStack() })
-            }
-
-            taggedComposable<Route.SettingsAppearance>(ScreenTags.SettingsAppearance) {
-                AppearanceSettingsScreen(
-                    onNavigateBack = { navController.popBackStack() }
-                )
-            }
-
-            taggedComposable<Route.SettingsLanguage>(ScreenTags.SettingsLanguage) {
-                LanguageScreen(
-                    onNavigateBack = { navController.popBackStack() }
-                )
-            }
-
-            taggedComposable<Route.SettingsLocation>(ScreenTags.SettingsLocation) {
-                LocationScreen(
-                    onNavigateBack = { navController.popBackStack() }
-                )
-            }
-
-            taggedComposable<Route.SettingsQuran>(ScreenTags.SettingsQuran) {
-                QuranSettingsScreen(
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToSelectReciter = { navController.navigate(Route.SelectReciter) },
-                    onNavigateToSelectTranslation = {
-                        navController.navigate(Route.SelectTranslation)
-                    }
-                )
-            }
-
-            taggedComposable<Route.SelectReciter>(ScreenTags.SelectReciter) {
-                SelectReciterScreen(
-                    onNavigateBack = { navController.popBackStack() }
-                )
-            }
-
-            taggedComposable<Route.SelectTranslation>(ScreenTags.SelectTranslation) {
-                SelectTranslationScreen(
-                    onNavigateBack = { navController.popBackStack() }
-                )
-            }
-
-            taggedComposable<Route.SettingsWidgets>(ScreenTags.SettingsWidgets) {
-                WidgetsScreen(
-                    onNavigateBack = { navController.popBackStack() }
-                )
-            }
-
-            taggedComposable<Route.SettingsAbout>(ScreenTags.SettingsAbout) {
-                val context = androidx.compose.ui.platform.LocalContext.current
-                val shareScope = rememberCoroutineScope()
-                var showShareSheet by remember { mutableStateOf(false) }
-                if (showShareSheet) {
-                    ShareAppSheet(
-                        onDismiss = { showShareSheet = false },
-                        onShareLink = {
-                            showShareSheet = false
-                            shareScope.launch {
-                                ContentShareManager.shareBranded(
-                                    context,
-                                    Shareables.appInvite(context),
-                                    includeText = true,
-                                )
-                            }
-                        },
-                    )
-                }
-                val contactEmail =
-                    stringResource(com.arshadshah.nimaz.core.ui.R.string.contact_email)
-                val contactSubject =
-                    stringResource(com.arshadshah.nimaz.core.ui.R.string.contact_email_subject)
-                AboutScreen(
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToPrivacyPolicy = {
-                        val intent = Intent(
-                            Intent.ACTION_VIEW,
-                            "https://nimaz.arshadshah.com/privacy".toUri()
-                        )
-                        context.startActivity(intent)
-                    },
-                    onNavigateToTerms = {
-                        val intent = Intent(
-                            Intent.ACTION_VIEW,
-                            "https://nimaz.arshadshah.com/terms".toUri()
-                        )
-                        context.startActivity(intent)
-                    },
-                    onNavigateToLicenses = {
-                        navController.navigate(Route.Licenses)
-                    },
-                    onRateApp = {
-                        val activity = context as? Activity
-                        if (activity != null) {
-                            val manager = ReviewManagerFactory.create(context)
-                            manager.requestReviewFlow().addOnCompleteListener { task ->
-                                if (task.isSuccessful) {
-                                    manager.launchReviewFlow(activity, task.result)
-                                } else {
-                                    val intent = Intent(
-                                        Intent.ACTION_VIEW,
-                                        "https://play.google.com/store/apps/details?id=com.arshadshah.nimaz".toUri()
-                                    )
-                                    context.startActivity(intent)
-                                }
-                            }
-                        }
-                    },
-                    onShareApp = { showShareSheet = true },
-                    onContactUs = {
-                        ContentShareManager.sendEmail(
-                            context,
-                            address = contactEmail,
-                            subject = contactSubject,
-                        )
-                    }
-                )
-            }
-
-            taggedComposable<Route.Licenses>(ScreenTags.Licenses) {
-                LicensesScreen(
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToDetail = { libraryId ->
-                        navController.navigate(Route.LicenseDetail(libraryId))
-                    }
-                )
-            }
-
-            taggedComposable<Route.LicenseDetail>(ScreenTags.LicenseDetail) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.LicenseDetail>()
-                LicenseDetailScreen(
-                    libraryId = args.libraryHashCode,
-                    onNavigateBack = { navController.popBackStack() }
-                )
-            }
-
-            // Names — three catalogues, one tabbed screen
-            taggedComposable<Route.Names>(ScreenTags.Names) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.Names>()
-                AdaptiveNamesScreen(
-                    onNavigate = { navController.navigate(it) },
-                    initialTab = NamesTab.fromOrdinal(args.tab),
-                    onNavigateBack = { navController.popBackStack() },
-                )
-            }
-
-            taggedComposable<Route.Favourites>(ScreenTags.Favourites) {
-                FavouritesScreen(
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToAsmaUlHusna = {
-                        navController.navigate(Route.AsmaUlHusnaDetail(it))
-                    },
-                    onNavigateToAsmaUnNabi = {
-                        navController.navigate(Route.AsmaUnNabiDetail(it))
-                    },
-                    onNavigateToProphet = { navController.navigate(Route.ProphetDetail(it)) },
-                )
-            }
-
-            taggedComposable<Route.AsmaUlHusnaDetail>(ScreenTags.AsmaUlHusnaDetail) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.AsmaUlHusnaDetail>()
-                AsmaUlHusnaDetailScreen(
-                    nameId = args.nameId,
-                    onNavigateBack = { navController.popBackStack() }
-                )
-            }
-
-            taggedComposable<Route.AsmaUnNabiDetail>(ScreenTags.AsmaUnNabiDetail) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.AsmaUnNabiDetail>()
-                AsmaUnNabiDetailScreen(
-                    nameId = args.nameId,
-                    onNavigateBack = { navController.popBackStack() }
-                )
-            }
-
-            taggedComposable<Route.ProphetDetail>(ScreenTags.ProphetDetail) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.ProphetDetail>()
-                ProphetDetailScreen(
-                    prophetId = args.prophetId,
-                    onNavigateBack = { navController.popBackStack() }
-                )
-            }
-
-            // Khatam screens
-            taggedComposable<Route.KhatamList>(ScreenTags.KhatamList) {
-                AdaptiveKhatamScreen(
-                    onNavigate = { navController.navigate(it) },
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToCreate = { navController.navigate(Route.KhatamCreate) },
-                )
-            }
-
-            taggedComposable<Route.KhatamDetail>(ScreenTags.KhatamDetail) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.KhatamDetail>()
-                KhatamDetailScreen(
-                    khatamId = args.khatamId,
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToRead = { surahNumber, ayahNumber ->
-                        navController.navigate(Route.QuranReader(surahNumber, ayahNumber))
-                    },
-                    onNavigateToEdit = { khatamId ->
-                        navController.navigate(Route.KhatamEdit(khatamId))
-                    }
-                )
-            }
-
-            // Create and edit share one screen; a null id means "create".
-            taggedComposable<Route.KhatamCreate>(ScreenTags.KhatamCreate) {
-                KhatamFormScreen(
-                    khatamId = null,
-                    onNavigateBack = { navController.popBackStack() }
-                )
-            }
-
-            taggedComposable<Route.KhatamEdit>(ScreenTags.KhatamEdit) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.KhatamEdit>()
-                KhatamFormScreen(
-                    khatamId = args.khatamId,
-                    onNavigateBack = { navController.popBackStack() }
-                )
-            }
-
-            taggedComposable<Route.SettingsHelp>(ScreenTags.SettingsHelp) {
-                val context = androidx.compose.ui.platform.LocalContext.current
-                // Unified support inbox — same address the About screen contacts.
-                val supportEmail =
-                    androidx.compose.ui.res.stringResource(com.arshadshah.nimaz.core.ui.R.string.contact_email)
-                val supportSubject =
-                    androidx.compose.ui.res.stringResource(com.arshadshah.nimaz.core.ui.R.string.nimaz_support_request)
-                com.arshadshah.nimaz.presentation.screens.help.HelpScreen(
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToTopic = { topicId ->
-                        navController.navigate(
-                            Route.HelpTopicDetail(
-                                topicId
-                            )
-                        )
-                    },
-                    onContact = {
-                        ContentShareManager.sendEmail(
-                            context,
-                            address = supportEmail,
-                            subject = supportSubject,
-                        )
-                    }
-                )
-            }
-
-            taggedComposable<Route.HelpTopicDetail>(ScreenTags.HelpTopicDetail) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.HelpTopicDetail>()
-                com.arshadshah.nimaz.presentation.screens.help.HelpTopicDetailScreen(
-                    topicId = args.topicId,
-                    onNavigateBack = { navController.popBackStack() },
-                    onOpenGuide = { guideId -> navController.navigate(Route.HelpGuide(guideId)) }
-                )
-            }
-
-            taggedComposable<Route.HelpGuide>(ScreenTags.HelpGuide) { backStackEntry ->
-                val args = backStackEntry.toRoute<Route.HelpGuide>()
-                com.arshadshah.nimaz.presentation.screens.help.HelpGuideScreen(
-                    guideId = args.guideId,
-                    onNavigateBack = { navController.popBackStack() },
-                    onDeepLink = { key ->
-                        helpDeepLinkRoute(key)?.let { route ->
-                            navController.navigate(route)
-                        }
-                    }
-                )
-            }
-
-            taggedComposable<Route.SettingsSync>(ScreenTags.SettingsSync) {
-                com.arshadshah.nimaz.presentation.screens.settings.SyncScreen(
-                    onNavigateBack = { navController.popBackStack() }
-                )
-            }
-
-            // Bookmarks
-            taggedComposable<Route.AllBookmarks>(ScreenTags.AllBookmarks) {
-                SavedScreen(
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToQuranAyah = { surah, ayah ->
-                        navController.navigate(Route.QuranReader(surah, ayah))
-                    },
-                    onNavigateToHadith = { bookId, hadithNumber ->
-                        navController.navigate(Route.HadithByNumber(bookId, hadithNumber))
-                    },
-                    onNavigateToDua = { duaId ->
-                        navController.navigate(Route.DuaReader(duaId))
-                    },
-                )
-            }
-
-            // Global Search
-            taggedComposable<Route.GlobalSearch>(ScreenTags.GlobalSearch) {
-                SearchScreen(
-                    enableAsk = true,
-                    onNavigateBack = { navController.popBackStack() },
-                    onNavigateToQuranAyah = { surah, ayah ->
-                        navController.navigate(Route.QuranReader(surah, ayah))
-                    },
-                    onNavigateToSurah = { surah ->
-                        navController.navigate(Route.QuranReader(surah))
-                    },
-                    onNavigateToHadith = { bookId, hadithId ->
-                        navController.navigate(Route.HadithReader(hadithId))
-                    },
-                    onNavigateToDua = { duaId ->
-                        navController.navigate(Route.DuaReader(duaId))
-                    },
-                    onNavigateToName = { catalog, id ->
-                        navController.navigate(
-                            when (catalog) {
-                                NameCatalog.ASMA_UL_HUSNA -> Route.AsmaUlHusnaDetail(id)
-                                NameCatalog.ASMA_UN_NABI -> Route.AsmaUnNabiDetail(id)
-                                NameCatalog.PROPHETS -> Route.ProphetDetail(id)
-                            }
-                        )
-                    },
-                    onNavigateToSearchSettings = { navController.navigate(Route.SearchSettings) },
-                    onNavigateToProof = { target -> navController.navigate(target.toRoute()) }
-                )
-            }
-
-            taggedComposable<Route.SearchSettings>(ScreenTags.SearchSettings) {
-                SearchSettingsScreen(
-                    onNavigateBack = { navController.popBackStack() }
-                )
-            }
+            // Every destination lives in its feature's graph extension. See #563: one file
+            // registering all 94 of them is what stopped any feature moving into its own
+            // module, because :app would have had to import from all eleven at once.
+            onboardingGraph(navController)
+            homeGraph(navController, analyticsContext)
+            quranGraph(navController)
+            contentGraph(navController)
+            trackerGraph(navController)
+            prayerGraph(navController)
+            calendarGraph(navController)
+            searchGraph(navController)
+            toolsGraph(navController)
+            aboutGraph(navController)
+            settingsGraph(navController)
         }
     }
 }
@@ -1411,7 +270,7 @@ private val bottomNavItems = listOf(
     BottomNavItem(Route.More, "More", Icons.Default.MoreHoriz)
 )
 
-private fun restartApp(context: Context) {
+internal fun restartApp(context: Context) {
     val intent = context.packageManager.getLaunchIntentForPackage(context.packageName)
     intent?.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP or Intent.FLAG_ACTIVITY_CLEAR_TASK or Intent.FLAG_ACTIVITY_NEW_TASK)
     context.startActivity(intent)
